Bring back feature-test-api
