+++ chain-impl-mockchain/Cargo.toml
@@ -37,9 +37,9 @@ quickcheck = "0.9"
 quickcheck_macros = "0.9"
 chain-core = { path = "../chain-core"}
 chain-test-utils = { path = "../chain-test-utils" }
-chain-crypto = { path = "../chain-crypto", features=["property-test-api"] }
-chain-time = { path = "../chain-time" }
-chain-addr = { path = "../chain-addr", features=["property-test-api"] }
+chain-crypto = { path = "../chain-crypto", features=["property-test-api"]}
+chain-time = { path = "../chain-time", features=["property-test-api"]}
+chain-addr = { path = "../chain-addr", features=["property-test-api"]}
 chain-storage = { path = "../chain-storage" }
 ed25519-bip32 = "0.3"
 rand_chacha = "0.2"
