+++ chain-impl-mockchain/src/ledger/ledger.rs
@@ -1620,11 +1620,10 @@ mod tests {
 
     impl Arbitrary for ArbitraryEmptyLedger {
         fn arbitrary<G: Gen>(g: &mut G) -> Self {
-            let mut r = chain_test_utils::utils::new_random_generator();
             let mut ledger = Ledger::empty(
                 Arbitrary::arbitrary(g),
                 Arbitrary::arbitrary(g),
-                chain_test_utils::time::generate_time_era(&mut r),
+                Arbitrary::arbitrary(g),
                 Arbitrary::arbitrary(g),
             );
