+++ chain-time/Cargo.toml
@@ -12,6 +12,10 @@ quickcheck_macros = { version = "0.9", optional = true }
 chain-ser = { path = "../chain-ser" }
 chain-core = { path = "../chain-core" }
 
+
+[features]
+property-test-api = ["quickcheck", "quickcheck_macros"]
+
 [dev-dependencies]
 quickcheck = "0.9"
 quickcheck_macros = "0.9"
